Updating displayAllProducts() to indicate added items

+++ inc/functions.php
@@ -140,10 +140,16 @@
                 echo "<td>" . $record["NAME"] . "</td>";
                 echo "<td>$" . $record["price"] . "</td>";
                 echo "<td>" . $record["description"] . "</td>";
-                echo "<form action='inc/addToCart.php'>";
-                echo "<input type='hidden' name='addProduct' value=" . $record['productId'] . " />";
-                echo "<td><input type='submit' class='btn btn-danger' value='Add to Cart'></td>";
-                echo "</form>";
+                
+                if(!array_key_exists($record['productId'], $_SESSION['cart'])){
+                    echo "<form action='inc/addToCart.php'>";
+                    echo "<input type='hidden' name='addProduct' value=" . $record['productId'] . " />";
+                    echo "<td><input type='submit' class='btn btn-primary btn-block' value='Add to Cart'></td>";
+                    echo "</form>";
+                } else {
+                    echo "<td><button type=button class='btn btn-secondary btn-block'>Added!</button></td>";
+                }
+                
                 echo "</tr>";
             }
             
@@ -173,10 +179,16 @@
                     echo "<td>" . $record["NAME"] . "</td>";
                     echo "<td>$" . $record["price"] . "</td>";
                     echo "<td>" . $record["description"] . "</td>";
-                    echo "<form action='inc/addToCart.php'>";
-                    echo "<input type='hidden' name='addProduct' value=" . $record['productId'] . " />";
-                    echo "<td><input type='submit' class='btn btn-danger' value='Add to Cart'></td>";
-                    echo "</form>";
+                    
+                    if(!array_key_exists($record['productId'], $_SESSION['cart'])){
+                        echo "<form action='inc/addToCart.php'>";
+                        echo "<input type='hidden' name='addProduct' value=" . $record['productId'] . " />";
+                        echo "<td><input type='submit' class='btn btn-primary btn-block' value='Add to Cart'></td>";
+                        echo "</form>";
+                    } else {
+                        echo "<td><button type=button class='btn btn-secondary btn-block'>Added!</button></td>";
+                    }
+                    
                     echo "</tr>";
                 }
